Fix all warnings

+++ algorithms/utils/euclidian_point.h
@@ -44,7 +44,7 @@ namespace parlayANN {
 
 float euclidian_distance_(const uint8_t *p, const uint8_t *q, unsigned d) {
   int result = 0;
-  for (int i = 0; i < d; i++) {
+  for (size_t i = 0; i < d; i++) {
     result += ((int32_t)((int16_t)q[i] - (int16_t)p[i])) *
       ((int32_t)((int16_t)q[i] - (int16_t)p[i]));
   }
@@ -53,7 +53,7 @@ float euclidian_distance_(const uint8_t *p, const uint8_t *q, unsigned d) {
 
 float euclidian_distance(const uint8_t *p, const uint8_t *q, unsigned d) {
   int32_t result = 0;
-  for (int i = 0; i < d; i++) {
+  for (size_t i = 0; i < d; i++) {
     int32_t qi = (int32_t) p[i];
     int32_t pi = (int32_t) q[i];
     result += (qi - pi) * (qi - pi);
@@ -63,7 +63,7 @@ float euclidian_distance(const uint8_t *p, const uint8_t *q, unsigned d) {
 
 float euclidian_distance(const uint16_t *p, const uint16_t *q, unsigned d) {
   int64_t result = 0;
-  for (int i = 0; i < d; i++) {
+  for (size_t i = 0; i < d; i++) {
     int32_t qi = (int32_t) p[i];
     int32_t pi = (int32_t) q[i];
     result += (qi - pi) * (qi - pi);
@@ -73,7 +73,7 @@ float euclidian_distance(const uint16_t *p, const uint16_t *q, unsigned d) {
 
 float euclidian_distance(const int8_t *p, const int8_t *q, unsigned d) {
   int result = 0;
-  for (int i = 0; i < d; i++) {
+  for (size_t i = 0; i < d; i++) {
     result += ((int32_t)((int16_t)q[i] - (int16_t)p[i])) *
       ((int32_t)((int16_t)q[i] - (int16_t)p[i]));
   }
@@ -84,7 +84,7 @@ float euclidian_distance(const float *p, const float *q, unsigned d) {
   //efanna2e::DistanceL2 distfunc;
   //return distfunc.compare(p, q, d);
   float result = 0.0;
-  for (int i = 0; i < d; i++)
+  for (size_t i = 0; i < d; i++)
     result += (q[i] - p[i]) * (q[i] - p[i]);
   return (float)result;
 }
